pyproject.toml
[project]
name = "dtfp-devtools"
version = "0.1.0"
description = "DTFP 파이프라인 통합 검증 및 데모 도구"
requires-python = ">=3.11"
dependencies = [
    "paho-mqtt>=2.0",
    "rich>=13.0",
    "streamlit>=1.30",
    "pytest>=8.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.ruff]
line-length = 100
target-version = "py311"
